Avoid unrolling loops with timing statements
Loops (like for or while) cannot be unrolled safely when
timing statements are a part of the loop. Hence, skip
unrolling in this case.

The update is from the suggestion of Krzysztof Bieganski
in the github issue, with the addition of checking if the global
timing option is True.

This fixes:
https://github.com/verilator/verilator/issues/4237